moved struct into cpp file, thanks Shreyas
